requirements.txt
absl_py==2.0.0
accelerate==0.32.1
deepeval==0.21.73
deepspeed==0.10.1+46d859a7
huggingface_hub==0.23.4
ipdb==0.13.13
joblib==1.3.2
langchain==0.2.14
lightgbm==4.5.0
matplotlib==3.9.2
numpy==1.24.4
openai==1.40.8
pandas==2.0.3
peft==0.12.0
promptsource==0.2.3
psutil==5.9.5
psutil==6.0.0
pytorch_lightning==2.0.2
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
rich==13.7.1
scikit_learn==1.3.0
scipy==1.14.0
seaborn==0.13.2
seqio==0.0.19
setuptools==71.0.4
setuptools==68.2.2
t5==0.9.4
tabpfn==0.1.10
tensorboard==2.15.1
tensorflow==2.17.0
tensorflow_datasets==4.9.6
torch==2.2.0
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
Wikipedia_API==0.6.0
xgboost==2.1.1
